Refactor integration tests and add SAML integrator integration test
